requirements.txt
arrow==1.3.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pandas==2.2.2
pip-autoremove==0.10.0
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
setuptools==72.1.0
six==1.16.0
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
